HTBHF-2521 move state machine to flow control directory
