+++ src/web/routes/application/common/confirmation-code.js
@@ -1,5 +1,5 @@
 const { CONFIRMATION_CODE_ENTERED_SESSION_PROPERTY } = require('./constants')
-const { stateMachine, states } = require('../common/state-machine')
+const { stateMachine, states } = require('../flow-control')
 
 // reset confirmation code and reset state to in progress so that the user is not taken straight back to check answers.
 const handleConfirmationCodeReset = req => {
