+++ src/web/routes/application/flow-control/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('./state-machine')
+}
