+++ src/web/routes/application/middleware/handle-post.test.js
@@ -123,7 +123,7 @@ test('handlePost() should invalidate review if required by step', (t) => {
   const INVALIDATE_REVIEW = 'INVALIDATE_REVIEW'
   const { handlePost } = proxyquire('./handle-post', {
     ...defaultValidator,
-    '../common/state-machine': {
+    '../flow-control': {
       stateMachine: { dispatch },
       actions: { INVALIDATE_REVIEW }
     }
