+++ src/web/routes/application/terms-and-conditions/post.js
@@ -6,7 +6,7 @@ const { validationResult } = require('express-validator')
 const { wrapError } = require('../common/formatters')
 const { logger } = require('../../../logger')
 const { REQUEST_ID_HEADER } = require('../../../server/headers')
-const { stateMachine, states, actions } = require('../common/state-machine')
+const { stateMachine, states, actions } = require('../flow-control')
 const { createRequestBody } = require('./create-request-body')
 const { isErrorStatusCode } = require('./predicates')
 const { CLAIMS_ENDPOINT, NO_ELIGIBILITY_STATUS_MESSAGE } = require('./constants')
